# docs/api/unified-api-spec.yaml
openapi: 3.0.3
info:
  title: SPARC Platform Unified API
  description: |
    Comprehensive API specification for the SPARC (Security Platform with Advanced Real-time Control) system.
    This unified specification covers all 24 microservices providing access control, video management, 
    environmental monitoring, visitor management, mobile credentials, and compliance features.
    
    ## Key Features
    - Multi-tenant architecture with tenant isolation
    - Real-time event processing and notifications
    - Offline resilience with 72-hour operation capability
    - Scalable to 10,000 doors and 1,000 concurrent video streams
    - SOX, HIPAA, and PCI-DSS compliance
    - Mobile credential management with mesh networking
    - Advanced analytics and reporting
    
    ## Authentication
    All endpoints require JWT authentication with tenant-specific scoping.
    Include the `Authorization: Bearer <token>` header and `X-Tenant-ID` header in all requests.
    
    ## Multi-Tenant Usage
    All API endpoints are tenant-scoped. Include the `X-Tenant-ID` header to specify the tenant context.
    Data isolation is enforced at the database and application levels.
  version: 1.0.0
  contact:
    name: SPARC Platform API Support
    email: api-support@sparc-platform.com
  license:
    name: Proprietary
    url: https://sparc-platform.com/license
servers:
  - url: https://api.sparc-platform.com/v1
    description: Production API Server
...
    description: Staging API Server
...
security:
  - BearerAuth: []
  - TenantHeader: []
paths:
...
  /elevators:
    get:
      tags:
        - Elevator Control Service
      summary: List elevators
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
        - name: building
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
            enum: [online, offline, maintenance, emergency, error]
        - name: manufacturer
          in: query
          schema:
            type: string
            enum: [otis, kone, schindler, thyssenkrupp, other]
      responses:
        '200':
          description: Elevators retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ElevatorListResponse'

    post:
      tags:
        - Elevator Control Service
      summary: Add new elevator
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateElevatorRequest'
      responses:
        '201':
          description: Elevator added successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Elevator'
  /elevators/{elevatorId}:
    get:
      tags:
        - Elevator Control Service
      summary: Get elevator details
      parameters:
        - name: elevatorId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Elevator details retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Elevator'

    put:
      tags:
        - Elevator Control Service
      summary: Update elevator configuration
      parameters:
        - name: elevatorId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateElevatorRequest'
      responses:
        '200':
          description: Elevator updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Elevator'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from /auth/login endpoint
    TenantHeader:
      type: apiKey
      in: header
      name: X-Tenant-ID
      description: Tenant identifier for multi-tenant isolation
  parameters:
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
      description: Page number for pagination
    LimitParam:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
      description: Number of items per page
...
  schemas:
...
    Elevator:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        building:
          type: string
        manufacturer:
          type: string
          enum: [otis, kone, schindler, thyssenkrupp, other]
        model:
          type: string
        capacity:
          type: integer
          description: Maximum capacity in kg
        floorsServed:
          type: array
          items:
            type: integer
        status:
          type: string
          enum: [online, offline, maintenance, emergency, error]
        currentFloor:
          type: integer
        direction:
          type: string
          enum: [up, down, idle]
        emergencyOverride:
          type: boolean
        lastMaintenance:
          type: string
          format: date-time
        nextMaintenance:
          type: string
          format: date-time
        configuration:
          type: object
          properties:
            speed:
              type: number
            doorOpenTime:
              type: integer
            maxWaitTime:
              type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    CreateElevatorRequest:
      type: object
      required:
        - name
        - building
        - manufacturer
        - floorsServed
      properties:
        name:
          type: string
        building:
          type: string
        manufacturer:
          type: string
          enum: [otis, kone, schindler, thyssenkrupp, other]
        model:
          type: string
        capacity:
          type: integer
        floorsServed:
          type: array
          items:
            type: integer
        configuration:
          type: object
    UpdateElevatorRequest:
      type: object
      properties:
        name:
          type: string
        capacity:
          type: integer
        floorsServed:
          type: array
          items:
            type: integer
        configuration:
          type: object
    ElevatorListResponse:
      type: object
      properties:
        elevators:
          type: array
          items:
            $ref: '#/components/schemas/Elevator'
        pagination:
          $ref: '#/components/schemas/PaginationInfo'
...
    PaginationInfo:
      type: object
      properties:
        page:
          type: integer
        limit:
          type: integer
        total:
          type: integer
        totalPages:
          type: integer
        hasNext:
          type: boolean
        hasPrevious:
          type: boolean
...